refactor: replace unwrap/expect with proper error handling
Core changes:
- Use OnceLock for static regex patterns in validation/rules.rs
- Use HeaderValue::from_static() for HTTP headers
- Fix CacheManager default impl with fallback
- Fix chrono::Duration conversion with proper error handling
- Fix vector access with Option handling
- Import Metric trait for reset() method
- Fix type mismatches in subagent executor and MCP scanner

Tools changes:
- Fix Tool trait imports in glob/write/json_edit modules
- Merge duplicate Tool impl for WriteTool
- Remove obsolete parameters_json_schema_legacy method
- Fix description() return type in diagnostic tools
- Remove duplicate patterns.rs file (keep directory module)

+++ crates/sage-core/src/agent/subagent/executor/executor.rs
@@ -171,7 +171,7 @@ impl SubAgentExecutor {
         duration_secs: f64,
     ) -> ExecutionMetadata {
         ExecutionMetadata {
-            total_tokens: total_usage.total_tokens,
+            total_tokens: total_usage.total_tokens as u64,
             total_tool_uses: tool_calls_count as u32,
             execution_time_ms: (duration_secs * 1000.0) as u64,
             tools_used: Vec::new(),
